Started a model implementation taking a more formalized approach.
The next step is to perhaps test that this works without the use of
`prevE`, then move the support `prevE`. That is where things get
interesting! We'll have to do some interesting case splitting as
described in my hadwritted notes. Lets hope I don't get hit by a bus
else future generations will be left to solve that mystery for
themselves. In essence, I'll have to case split on an unknown time for
the first event occurence in an time span... something like that.