I got confused at this point and tried to remove the "clearance" logic.
My trail of thought was that as is described below. The problem is that
I tried to justify removing necessary/usefully logic (clearance logic) from the
theory layer by arguing that the higher level interface makes more
garantees that means we don't need the clearance logic. Since the
clearance logic is not wrong and since the higher level interface's
design is still a WIP, getting rid of clearance is just a dumb stupid
idea!

------- Old Comment -------

WIP Clearance checks are not actually necessary

Dispite much fuss about calculating clearance, this turns out not to be
necessary! The motivating example for the use of clearnace checks was
a (transitive) self reference. The increment test case is the simplest
example:

  incE :: EIx ()
  a = do
    aPrevVal <- fromMaybe 0 <$> prevV a
    e <- getE incE
    case e of
      NoOcc -> aPrevVal
      Occ () -> aPrevVal + 1

The problem is that the derivation is immediatelly stuck waiting on a
self reference which will never produce Facts.

Why is this now resolved?!?!?!?